Adding a simulation
This code has two main methods: buildseq and periodic. buildseq takes in four inputs, the first two terms of the sequence, the function f(y) and the number of iterations, it outputs the recurrent sequence up to n terms as a list. periodic takes in one input a list and outputs the period of the list, if the list is not periodic it returns false. There are also some basic functions f(y).